Refactoring stubs.
