+++ tests/StudentToolkit.Tests/GlobalUsings.cs
@@ -5,9 +5,9 @@ global using StudentToolkit.MVVM.Models.Navigation.Messages;
 global using StudentToolkit.MVVM.Models.Navigation.Messages.Queries;
 global using StudentToolkit.MVVM.ViewModels;
 global using StudentToolkit.MVVM.ViewModels.Base;
-global using StudentToolkit.Tests.Stubs.Messages;
-global using StudentToolkit.Tests.Stubs.Messages.Queris;
-global using StudentToolkit.Tests.Stubs.ViewModels;
+global using StudentToolkit.Tests.Stubs.Wpf.MVVM.Models.Messages;
+global using StudentToolkit.Tests.Stubs.Wpf.MVVM.Models.Messages.Queris;
+global using StudentToolkit.Tests.Stubs.Wpf.MVVM.ViewModels;
 global using StudentToolkit.WpfCore.Exceptions.Navigation;
 global using StudentToolkit.WpfCore.Services;
