--- tests/StudentToolkit.Tests/Stubs/DataTemplate/StubTwoViewModel.cs
@@ -1,3 +1,0 @@
-﻿namespace StudentToolkit.Tests.Stubs.DataTemplate;
-
-internal class StubTwoViewModel { }
