--- tests/StudentToolkit.Tests/Stubs/Messages/StubNavigationMessage.cs
@@ -1,7 +1,0 @@
-﻿namespace StudentToolkit.Tests.Stubs.Messages;
-
-public sealed class StubNavigationMessage : ValueChangedMessage<NavigationModel>
-{
-    public StubNavigationMessage(NavigationModel value) 
-        : base(value) { }
-}
