--- tests/StudentToolkit.Tests/Stubs/ViewModels/StubNotNavigatingViewModel.cs
@@ -1,3 +1,0 @@
-﻿namespace StudentToolkit.Tests.Stubs.ViewModels;
-
-public class StubNotNavigatingViewModel : ViewModel { }
