+++ tests/StudentToolkit.Tests/Stubs/Wpf/MVVM/Models/Messages/StubNavigationMessage.cs
@@ -1,0 +1,7 @@
+﻿namespace StudentToolkit.Tests.Stubs.Wpf.MVVM.Models.Messages;
+
+public sealed class StubNavigationMessage : ValueChangedMessage<NavigationModel>
+{
+    public StubNavigationMessage(NavigationModel value)
+        : base(value) { }
+}
