+++ tests/StudentToolkit.Tests/Stubs/Wpf/MVVM/ViewModels/StubNavigatingViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace StudentToolkit.Tests.Stubs.ViewModels;
+﻿namespace StudentToolkit.Tests.Stubs.Wpf.MVVM.ViewModels;
 
 /// <summary>
 /// Stub view model for testing navigation.
