+++ tests/StudentToolkit.Tests/Stubs/Wpf/MVVM/ViewModels/StubNotNavigatingViewModel.cs
@@ -1,0 +1,3 @@
+﻿namespace StudentToolkit.Tests.Stubs.Wpf.MVVM.ViewModels;
+
+public class StubNotNavigatingViewModel : ViewModel { }
