+++ tests/StudentToolkit.Tests/Stubs/Wpf/MVVM/ViewModels/StubTwoViewModel.cs
@@ -1,0 +1,3 @@
+﻿namespace StudentToolkit.Tests.Stubs.Wpf.MVVM.ViewModels;
+
+internal class StubTwoViewModel { }
