+++ tests/StudentToolkit.Tests/Stubs/Wpf/MVVM/Views/StubTwoView.cs
@@ -1,5 +1,5 @@
 ﻿using System.Windows;
 
-namespace StudentToolkit.Tests.Stubs.DataTemplate;
+namespace StudentToolkit.Tests.Stubs.Wpf.MVVM.Views;
 
 internal class StubTwoView : FrameworkElement { }
